refactor: RegionDropdown 컴포넌트 분리 (#22)
